% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{create_epic}
\alias{create_epic}
\title{create epic}
\usage{
create_epic(
  name,
  created_at = Sys.time(),
  deadline = NULL,
  description = NULL,
  epic_state_id = get_all("epic-workflow")$default_epic_state_id,
  milestone_id = NA,
  planned_start_date = NA,
  owner_ids = NULL,
  group_ids = NULL,
  requested_by_id = NULL,
  config = list(),
  sc_token = get_token()
)
}
\arguments{
\item{name}{Required. The Epic’s name.}

\item{created_at}{Time created. Default is \code{Sys.time()}}

\item{deadline}{POSIXct format. The Epic’s deadline. Default NULL}

\item{description}{chr The Epic’s description. Default NULL}

\item{epic_state_id}{int The ID of the Epic State. See
\code{get_all("epic-workflow")}, Defaults to your accounts default state}

\item{milestone_id}{int The ID of the Milestone this Epic is related to. Default NULL}

\item{planned_start_date}{POSIXct. The Epic’s planned start date. Default NULL}

\item{owner_ids}{An array of UUIDs for any members you want to add as Owners on this new Epic.}

\item{group_ids}{An array of UUIDS for Groups to which this Epic is related.}

\item{requested_by_id}{uuid The ID of the member that requested the epic. Default NULL}

\item{config}{(optional) additional configuration to add to header}

\item{sc_token}{shortcut API token.  Defaults to \code{get_token()}}
}
\description{
create epic
}
\examples{
\dontrun{
create_epic( name = "foobar",
             deadline = as.POSIXct("2022-06-30 17:00:00"),
             description = "this is a test"
             )
}
}
