% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{create_iteration}
\alias{create_iteration}
\title{create iteration}
\usage{
create_iteration(
  name,
  start_date,
  end_date,
  description,
  follower_ids = NA,
  group_ids = NA,
  labels = NA,
  config = list(),
  sc_token = get_token()
)
}
\arguments{
\item{name}{Required. The name of this Iteration.}

\item{start_date}{as character. Required. The date this Iteration begins, e.g. "2019-07-01".}

\item{end_date}{as character. Required. The date this Iteration ends, e.g. "2019-07-01".}

\item{description}{Required. The description of the Iteration.}

\item{follower_ids}{TODO An array of UUIDs for any Members you want to add as Followers.}

\item{group_ids}{TODO An array of UUIDs for any Groups you want to add as Followers. Currently, only one Group association is presented in our web UI.}

\item{labels}{An array of Labels attached to the Iteration.}
}
\description{
create iteration
}
\examples{
\dontrun{
create_iteration(
  name = "09 Jan - 15 Jan 2022",
  description = "Sprint Q1-1 2022",
  start_date = "2022-01-09",
  end_date = "2022-01-15"
)
}

}
