% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{delete_iteration}
\alias{delete_iteration}
\title{delete iteration}
\usage{
delete_iteration(iteration_public_id, config = list(), sc_token = get_token())
}
\arguments{
\item{iteration_public_id}{Required. The unique ID of the Iteration.}

\item{config}{(optional) additional configuration to add to header}

\item{sc_token}{shortcut API token.  Defaults to \code{get_token()}}
}
\description{
delete iteration
}
