% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{list_epic_stories}
\alias{list_epic_stories}
\title{List Epic Stories}
\usage{
list_epic_stories(epic_id, config = list(), sc_token = get_token())
}
\arguments{
\item{epic_id}{Epic Id as numeric / character string}

\item{config}{additional configuration to add to header}

\item{sc_token}{shortcut API token}
}
\value{
Data frame with stories belonging to that Epic
}
\description{
List Epic Stories
}
\examples{
\dontrun{
list_epic_stories(123)
}
}
