% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sc_get_one}
\alias{sc_get_one}
\title{Get endpoint data for one item}
\usage{
sc_get_one(
  id,
  endpoint = NULL,
  url = NULL,
  config = list(),
  sc_token = get_token()
)
}
\arguments{
\item{id}{the id for the record you want to retrieve}

\item{endpoint}{data to list. One of "categories", "entity-templates", "epic-workflows", "epics",
"files", "groups", "iterations"  ,"labels", "linked-files",
"members", "milestones", "projects", "repositories", "stories",
"story-links"}

\item{url}{shortcut API url.  Defaults to API V2}

\item{config}{additional configuration to add to header}

\item{sc_token}{shortcut API token.}
}
\description{
Get endpoint data for one item
}
\examples{
\dontrun{
# Retrieve all projects
sc_list_all(endpoint = "projects")

# List epics then retrieve full details about the first one
df <- sc_list_all("epics", response_type = "minimal")
sc_get_one("epics", id = df[ 1, "id"] )
}

}
