% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{sc_list_all}
\alias{sc_list_all}
\title{Get endpoint data}
\usage{
sc_list_all(
  endpoint = NULL,
  url = NULL,
  config = list(),
  sc_token = get_token(),
  response_type = "full"
)
}
\arguments{
\item{endpoint}{data to list. One of "categories", "entity-templates","epic-workflow", "epics"
,"files", "groups", "iterations","labels", "linked-files", "members",
"milestones", "projects", "repositories",
"workflows"}

\item{url}{shortcut API url.  Defaults to API V2}

\item{config}{additional configuration to add to header}

\item{sc_token}{shortcut API token.}

\item{response_type}{either "full" (all data) or "minimal" (just "entity_type", "id",
"name", "description"). Defaults to "full"}
}
\description{
Get endpoint data
}
\examples{

\dontrun{

# Retrieve all projects
sc_list_all(endpoint = "projects")

# List epics then retrieve full details about the first one
df <- sc_list_all("epics", response_type = "minimal")
sc_get_one("epics", id = df[ 1, "id"] )
}

}
