% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{sc_searsc_stories}
\alias{sc_searsc_stories}
\title{Search stories}
\usage{
sc_searsc_stories(
  query,
  searsc_limit = 1000,
  url = NULL,
  config = list(),
  response_type = "summary",
  sc_token = get_token()
)
}
\arguments{
\item{query}{Search query.  See https://help.shortcut.io/hc/en-us/articles/115005967026 for more detail on allowed search structure.}

\item{searsc_limit}{Maximum records to retrieve.  Defaults to 1000.}

\item{url}{Alternatively specify query as a url string. Useful if you want to use a different version of the API (default is V2)}

\item{config}{additional configuration to add to header}

\item{response_type}{either "full" (all data as a nested list) or
"summary" (flattened data frame with nested lists removed) or
"minimal" (just "entity_type", "id", "name", "description"). Defaults to "summary"}

\item{sc_token}{shortcut API token.}
}
\description{
Search stories
}
\examples{
\dontrun{
# Search for the term "foo" anywhere in stories
sc_searsc_stories(query = "foo")

# Search for epic names that contain specified text
# Note - no whitespace areound the ":"
sc_searsc_stories(query = "epic:foo")

# Combine search terms with "AND" ("OR" combinations are not supported by the API V2)
# Note the whitespace between each filter
sc_searsc_stories(query = "epic:foo is:done type:bug")


}

}
