% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{sc_url}
\alias{sc_url}
\title{shortcut url}
\usage{
sc_url(
  sc_base_url = get_url(),
  endpoint = NULL,
  id = NULL,
  sc_token = get_token(),
  query = NULL
)
}
\arguments{
\item{sc_base_url}{the base url. Defaults to V2 of API.}

\item{endpoint}{endpoint to retrieve (e.g. project, team, story)}

\item{id}{id of single record to be retrieved using \code{sc_get_one} function}

\item{sc_token}{shortcut API token.}

\item{query}{optional query search term}
}
\description{
Function to construct urls for shortcut api calls
}
\details{
This function constructs the url for each API call. \code{sc_token}
may be input as function parameters or set once with \code{set_token}.
}
