% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{create_iteration}
\alias{create_iteration}
\title{create iteration}
\usage{
create_iteration(
  name,
  start_date,
  end_date,
  description,
  config = list(),
  sc_token = get_token()
)
}
\arguments{
\item{name}{Required. The name of this Iteration.}

\item{start_date}{as character. Required. The date this Iteration begins, e.g. "2019-07-01".}

\item{end_date}{as character. Required. The date this Iteration ends, e.g. "2019-07-01".}

\item{description}{Required. The description of the Iteration.}

\item{config}{(optional) additional configuration to add to header}

\item{sc_token}{shortcut API token.  Defaults to \code{get_token()}}
}
\description{
create iteration
}
\details{
arguments for \code{follower_ids}, \code{group_ids}, \code{labels} are not yet implemented. See \href{https://github.com/techisdead/shortcutR/issues/2}{issue 2} to contribute.
}
\examples{
\dontrun{
create_iteration(
  name = "09 Jan - 15 Jan 2022",
  description = "Sprint Q1-1 2022",
  start_date = "2022-01-09",
  end_date = "2022-01-15"
)
}

}
